+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.0.8
+
+* Fix to be able to scroll inside the drawer 
+* Change Property `isswipeable` to `swipeable`
+* Change Property `isinitcreate` to `initcreate`
+* Change Property `ishistory` to `history`
+* Change default value of `maxsize` property to unspecified(-1)
+
 ## 0.0.4
 
 * Fix history back when touching an overlay element during closed animation.